Convert SkinNum to UINT16 and raise skin limit to 512

+++ src/d_clisrv.c
@@ -1068,7 +1068,7 @@ static void SV_SendPlayerInfo(INT32 node)
 
 		netbuffer->u.playerinfo[i].score = LONG(players[i].score);
 		netbuffer->u.playerinfo[i].timeinserver = SHORT((UINT16)(players[i].jointime / TICRATE));
-		netbuffer->u.playerinfo[i].skin = (UINT8)players[i].skin;
+		netbuffer->u.playerinfo[i].skin = (UINT16)players[i].skin;
 
 		// Extra data
 		// Kart has extra skincolors, so we can't use this
